Started to check for ray hit

+++ files/includes/miniRT_struct.h
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   structs.h                                          :+:      :+:    :+:   */
+/*   miniRT_struct.h                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: dkeraudr <dkeraudr@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/16 19:40:18 by dkeraudr          #+#    #+#             */
-/*   Updated: 2024/01/17 22:07:02 by dkeraudr         ###   ########.fr       */
+/*   Updated: 2024/01/28 18:55:05 by dkeraudr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -56,7 +56,7 @@ typedef struct s_hittable {
 typedef struct s_camera {
 	t_point		center;
 	t_vector	orientation;
-	int			fov;
+	double			fov;
 }				t_camera;
 
 typedef struct s_light {
